server/app/controllers/game/game.controller.spec.ts
import { CreateGameDto } from '@app/model/dto/game/create-game.dto';
import { GameService } from '@app/services/game/game.service';
import { GAMES_STUB } from '@app/tests.support/stubs/games.stub';
import { NEW_GAME_STUB } from '@app/tests.support/stubs/new-game.stub';
import { HttpStatus, INestApplication, Logger } from '@nestjs/common';
import { Test, TestingModule } from '@nestjs/testing';
import { SinonStubbedInstance, createStubInstance } from 'sinon';
import * as request from 'supertest';
import { GameController } from './game.controller';

describe('GameController', () => {
    let controller: GameController;
    let gameService: SinonStubbedInstance<GameService>;
    let logger: SinonStubbedInstance<Logger>;
    let app: INestApplication;

    beforeEach(async () => {
        gameService = createStubInstance(GameService);
        logger = createStubInstance(Logger);

        const module: TestingModule = await Test.createTestingModule({
            controllers: [GameController],
            providers: [
                { provide: GameService, useValue: gameService },
                { provide: Logger, useValue: logger },
            ],
        }).compile();
        app = module.createNestApplication();
        await app.init();
        controller = module.get<GameController>(GameController);
    });
    afterEach(async () => {
        await app.close();
    });

    it('should be defined', () => {
        expect(controller).toBeDefined();
    });

    // Tests des routes
    describe('GET /games', () => {
        it('should return all games', async () => {
            jest.spyOn(gameService, 'getAllGames').mockResolvedValue(Object.assign(GAMES_STUB()));

            const response = await request(app.getHttpServer()).get('/games').expect(HttpStatus.OK);

            expect(response.body).toEqual(Object.assign(GAMES_STUB()));
            expect(response.status).toBe(HttpStatus.OK);
        });

        it('should handle errors and return 404 status', async () => {
            const errorMessage = 'An error occurred';
            jest.spyOn(gameService, 'getAllGames').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer()).get('/games').expect(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
            expect(response.status).toBe(HttpStatus.NOT_FOUND);
        });
    });

    describe('POST /title', () => {
        it('should return title existence with status 200', async () => {
            const fakeTitle = { title: 'existing title' };
            const titleExistence = { titleExists: true };

            jest.spyOn(gameService, 'titleExists').mockResolvedValue(titleExistence.titleExists);

            const response = await request(app.getHttpServer())
                .post('/games/title')
                .send(fakeTitle)
                .set('Accept', 'application/json')
                .expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
            expect(response.body).toEqual(titleExistence);
        });

        it('should handle missing title and return status 200', async () => {
            const fakeTitle = { title: 'Non-existing Title' };
            const titleExistence = { titleExists: false };

            jest.spyOn(gameService, 'titleExists').mockResolvedValue(titleExistence.titleExists);

            const response = await request(app.getHttpServer())
                .post('/games/title')
                .send(fakeTitle)
                .set('Accept', 'application/json')
                .expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
            expect(response.body).toEqual(titleExistence);
        });

        it('should handle errors and return status 404', async () => {
            const errorMessage = 'An error occurred';
            const fakeTitle = { title: 'Title' };

            jest.spyOn(gameService, 'titleExists').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer())
                .post('/games/title')
                .send(fakeTitle)
                .set('Accept', 'application/json')
                .expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });

    describe('GET /:id', () => {
        it('should return a game with status 200', async () => {
            jest.spyOn(gameService, 'getGameById').mockResolvedValue(Object.assign(GAMES_STUB())[0]);

            const response = await request(app.getHttpServer()).get(`/games/${GAMES_STUB()[0].id}`).expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
            expect(response.body).toEqual(GAMES_STUB()[0]);
        });

        it('should handle game not found and return status 404', async () => {
            const fakeGame = null;
            jest.spyOn(gameService, 'getGameById').mockResolvedValue(fakeGame);

            const response = await request(app.getHttpServer()).get('/games/inexistentId').expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain('Jeu introuvable');
        });

        it('should handle errors and return status 404', async () => {
            const errorMessage = 'An error occurred';
            jest.spyOn(gameService, 'getGameById').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer()).get(`/games/${GAMES_STUB()[0].id}`).expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });

    describe('POST /', () => {
        const fakeGame: CreateGameDto = Object.assign(NEW_GAME_STUB());
        it('should add a new game with status 200', async () => {
            jest.spyOn(gameService, 'getGameById').mockResolvedValue(null);
            jest.spyOn(gameService, 'titleExists').mockResolvedValue(false);

            const response = await request(app.getHttpServer()).post('/games').send(fakeGame).set('Accept', 'application/json').expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
            expect(response.body).toEqual(fakeGame);
        });

        it('should handle conflict and return status 409 when game with same title already exists', async () => {
            jest.spyOn(gameService, 'getGameById').mockResolvedValue(null);
            jest.spyOn(gameService, 'titleExists').mockResolvedValue(true);

            const response = await request(app.getHttpServer())
                .post('/games')
                .send(fakeGame)
                .set('Accept', 'application/json')
                .expect(HttpStatus.CONFLICT);

            expect(response.status).toBe(HttpStatus.CONFLICT);
            expect(response.text).toContain('This game already exists');
        });

        it('should handle conflict and return status 409 when game already exists', async () => {
            jest.spyOn(gameService, 'getGameById').mockResolvedValue(fakeGame);

            const response = await request(app.getHttpServer())
                .post('/games')
                .send(fakeGame)
                .set('Accept', 'application/json')
                .expect(HttpStatus.CONFLICT);

            expect(response.status).toBe(HttpStatus.CONFLICT);
            expect(response.text).toContain('This game already exists');
        });
        it('should handle errors and return status 404', async () => {
            const errorMessage = 'An error occurred';
            jest.spyOn(gameService, 'getGameById').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer())
                .post('/games')
                .send(fakeGame)
                .set('Accept', 'application/json')
                .expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });

    describe('PUT /', () => {
        it('should update a game with status 200', async () => {
            jest.spyOn(gameService, 'updateGame').mockResolvedValue(null);

            const response = await request(app.getHttpServer())
                .put('/games')
                .send(Object.assign(GAMES_STUB())[0])
                .set('Accept', 'application/json')
                .expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
        });

        it('should handle errors and return status 404', async () => {
            const errorMessage = 'An error occurred';
            jest.spyOn(gameService, 'updateGame').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer())
                .put('/games')
                .send(Object.assign(GAMES_STUB())[0])
                .set('Accept', 'application/json')
                .expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });

    describe('DELETE /:id', () => {
        it('should delete a game with status 200', async () => {
            const gameId = Object.assign(GAMES_STUB())[0].id;
            jest.spyOn(gameService, 'deleteGameById').mockResolvedValue(null);

            const response = await request(app.getHttpServer()).delete(`/games/${gameId}`).expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
        });

        it('should handle errors and return status 404', async () => {
            const gameId = 'errorId';
            const errorMessage = 'An error occurred';

            jest.spyOn(gameService, 'deleteGameById').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer()).delete(`/games/${gameId}`).expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });

    describe('PATCH /:id/update-visibility', () => {
        it('should update game visibility with status 200', async () => {
            const gameId = Object.assign(GAMES_STUB())[0].id;
            const updateData = { isVisible: true };

            jest.spyOn(gameService, 'updateGameVisibility').mockResolvedValue(null);

            const response = await request(app.getHttpServer())
                .patch(`/games/${gameId}/update-visibility`)
                .send(updateData)
                .set('Accept', 'application/json')
                .expect(HttpStatus.OK);

            expect(response.status).toBe(HttpStatus.OK);
        });

        it('should handle errors and return status 404', async () => {
            const gameId = 'errorId';
            const updateData = { isVisible: true };
            const errorMessage = 'An error occurred';

            jest.spyOn(gameService, 'updateGameVisibility').mockRejectedValue(new Error(errorMessage));

            const response = await request(app.getHttpServer())
                .patch(`/games/${gameId}/update-visibility`)
                .send(updateData)
                .set('Accept', 'application/json')
                .expect(HttpStatus.NOT_FOUND);

            expect(response.status).toBe(HttpStatus.NOT_FOUND);
            expect(response.text).toContain(errorMessage);
        });
    });
});
